♻️ refactor : 요소 클릭 후 앞면뒷면 변경시 선택 해제 #124
