+++ src/components/editor/elements/text/text-sidebar.tsx
@@ -3,7 +3,7 @@ import { TextElement, useEditorStore } from '@/store/editor.store';
 import React from 'react';
 import { v4 as uuidv4 } from 'uuid';
 import { sideBarStore } from '@/store/editor.sidebar.store';
-import { calculateToolbarPosition } from '@/utils/editor/editor-cal-toolbar-position';
+import { calculateToolbarPosition } from '@/utils/editor/editor-calculate-toolbar-position';
 
 const TEXT_PRESETS = {
   TITLE: {
