+++ src/components/editor/elements/uploads/uploads-sidebar.tsx
@@ -3,7 +3,7 @@ import React, { ChangeEvent, useEffect, useState } from 'react';
 import { v4 as uuidv4 } from 'uuid';
 import { useEditorStore, UploadElement } from '@/store/editor.store';
 import { sideBarStore } from '@/store/editor.sidebar.store';
-import { calculateToolbarPosition } from '@/utils/editor/editor-cal-toolbar-position';
+import { calculateToolbarPosition } from '@/utils/editor/editor-calculate-toolbar-position';
 
 interface UploadedFile {
   id: string;
