+++ src/utils/editor/editor-calculate-toolbar-position.ts
